Externalize parameterized tests to new file format
This is so I can eventually convert the nose test file into a one and done kind of thing, but still have a nice and concise test description language.  It also allows that list of tests to eventually be created by an external entity (e.g. website)

Still have some more work to do to before finishing #4 though.