Refactor parser function calls
It wasn't immediately clear what some of the variables were and why there was somuch put in to the parser init calls, so I broke the different fuctions up much more clearly. i.e. you should be able to set the context to PRS and conjugate as much as you'd like, without having to re-set the context every time you want to inflect something.

This change also add the language header to the tests.yufa files -- now the location of which can be agnostic, as long as the grammar definitions follow the pattern of {language}/{inflection_type}.tx.