Add default inflection group and Minus ability for inflections
Doing pattern matching is probably not going to work well in the inflection level, so I'm going to just use the Set/If setup that exists in the normal ruleset for now.

Also remembered I had Quantity in there for a reason: to make sure the Nx(whatever) is the same, rather than two random (whatevers).